airbag synced

+++ multiple_video.py
@@ -13,9 +13,11 @@ DEBUG_MODE = args['debug']
 
 def move_video(folder_input, folder_output, dst_fd='data_const/run', result_df='data_const/final_vis',
                debug_mode=DEBUG_MODE):
-    shutil.rmtree('data_heavy', ignore_errors=True)
-    shutil.rmtree('sph_data/mc_solutions', ignore_errors=True)
-    shutil.rmtree('sph_data/mc_solutions_smoothed', ignore_errors=True)
+
+    if not debug_mode:
+        shutil.rmtree('data_heavy', ignore_errors=True)
+        shutil.rmtree('sph_data/mc_solutions', ignore_errors=True)
+        shutil.rmtree('sph_data/mc_solutions_smoothed', ignore_errors=True)
 
     sub_folders = os.walk(folder_input).__next__()[1]
     sub_out_folders = os.walk(folder_output).__next__()[1]
